place order is added

+++ src/main/java/com/jnkziaa/customerordercasestudy/dto/ResponseOrderDTO.java
@@ -13,5 +13,5 @@ public class ResponseOrderDTO {
     private int invoiceNumber;
     private String date;
     private String orderDescription;
-    private int orderID;
+    private Long orderID;
 }
